backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
python-pptx>=0.6.23
pdfplumber>=0.11.0
pypdf>=4.0.0
requests>=2.31.0
